+++ event.go
@@ -12,7 +12,7 @@ type Event struct {
 	Email     string                 `json:"email,omitempty"`
 	UserID    string                 `json:"user_id,omitempty"`
 	EventName string                 `json:"event_name,omitempty"`
-	CreatedAt int32                  `json:"created_at,omitempty"`
+	CreatedAt int64                  `json:"created_at,omitempty"`
 	Metadata  map[string]interface{} `json:"metadata,omitempty"`
 }
